Linux packaging script will use tar if 7zip not available.
The package.sh script now performs a check to see if 7z is a
valid shell command. If it's not, the script will use tar instead.
Also added some minor tweaks to check for the presence of files
before deleting them.

+++ deployment/package.sh
@@ -10,11 +10,11 @@ version_folder="${ta_folder}_${version}"
 cd "${0%/*}"
 
 # Clean up any existing packages for this version.
-rm /tmp/$version_folder.spl --force
-rm ../packages/$version_folder.spl --force
+if [ -x /tmp/$version_folder.spl ]; then rm /tmp/$version_folder.spl --force; fi
+if [ -x ../packages/$version_folder.spl ]; then rm ../packages/$version_folder.spl --force; fi
 
 # Create folder structure and file contents to put in package.
-rm $ta_folder --force --recursive
+if [ -x $ta_folder ]; then rm $ta_folder --force --recursive; fi
 mkdir $ta_folder 
 cp -r ../bin $ta_folder
 cp -r ../default $ta_folder
@@ -24,12 +24,15 @@ cp ../LICENSE $ta_folder
 cp ../README.md $ta_folder
 
 # Build the package file for this version
-7z a -ttar /tmp/$version_folder.tar $ta_folder/*
-cp /tmp/$version_folder.tar /tmp/$version_folder.spl
-7z a -tgzip ../packages/$version_folder.spl /tmp/$version_folder.spl
-
-# Clean up temporary working files used for packaging
-rm /tmp/$version_folder.spl
-rm /tmp/$version_folder.tar
+if command -v 7z 2>/dev/null; then
+  7z a -ttar /tmp/$version_folder.tar $ta_folder/*
+  cp /tmp/$version_folder.tar /tmp/$version_folder.spl
+  7z a -tgzip ../packages/$version_folder.spl /tmp/$version_folder.spl
+  # Clean up temporary working files used for packaging
+  rm /tmp/$version_folder.tar
+  rm /tmp/$version_folder.spl
+else
+  tar czf ../packages/$version_folder.spl $ta_folder/
+fi
 
 rm $ta_folder --force --recursive
